requirements.txt
einops
requests
pytest
tqdm
ipykernel
notebook
datasets
ipywidgets==7.7.1
jupyterlab-widgets==1.1.1
jupyter
matplotlib
numpy-stl
mediapy
pandas
scikit-learn
wandb==0.13.10
ftfy
tensorboard
plotly
sympy
huggingface_hub
gradio
accelerate
pre-commit
diffusers
gdown
jaxtyping
tiktoken
typeguard
torchinfo
torchvision
frozendict
graphviz
openai
transformer_lens==1.17.0
autorom[accept-rom-license]
ale-py
protobuf==3.20.3
imageio-ffmpeg
nnsight
torchtext
portalocker>=2.0.0
git+https://github.com/callummcdougall/CircuitsVis.git#subdirectory=python
git+https://github.com/callummcdougall/eindex.git
git+https://github.com/neelnanda-io/neel-plotly
